requirements.txt
# Core Django
Django==4.2.25
gunicorn==21.2.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Image Processing
Pillow==11.3.0
numpy==2.0.2

# AWS S3
boto3==1.40.55
django-storages==1.14.2

# Caching
redis==5.0.1
django-redis==5.4.0

# Email
django-anymail==10.2

# Security
django-cors-headers==4.3.1
django-environ==0.11.2

# Monitoring
sentry-sdk==1.38.0

# Development
django-debug-toolbar==4.2.0
